refactor: update placeholder package structure and usage
