+++ API/src/main/java/fr/maxlego08/menu/api/itemstack/components/LoreComponent.java
@@ -2,6 +2,7 @@ package fr.maxlego08.menu.api.itemstack.components;
 
 import fr.maxlego08.menu.api.context.BuildContext;
 import fr.maxlego08.menu.api.itemstack.ItemComponent;
+import fr.maxlego08.menu.api.placeholder.Placeholder;
 import org.bukkit.entity.Player;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.meta.ItemMeta;
@@ -27,7 +28,9 @@ public class LoreComponent extends ItemComponent {
     public void apply(@NotNull BuildContext context, @NotNull ItemStack itemStack, @Nullable Player player) {
         ItemMeta itemMeta = itemStack.getItemMeta();
         if (itemMeta != null) {
-            itemMeta.setLore(this.lore);
+
+            itemMeta.setLore(Placeholder.Placeholders.getPlaceholder().setPlaceholders(player,context.getPlaceholders().parse(this.lore)));
+
             itemStack.setItemMeta(itemMeta);
         }
     }
