+++ API/src/main/java/fr/maxlego08/menu/api/placeholder/AutoPlaceholder.java
@@ -1,4 +1,4 @@
-package fr.maxlego08.menu.placeholder;
+package fr.maxlego08.menu.api.placeholder;
 
 import fr.maxlego08.menu.api.interfaces.ReturnBiConsumer;
 import org.bukkit.OfflinePlayer;
