+++ Common/src/main/java/fr/maxlego08/menu/common/utils/PapiUtils.java
@@ -1,7 +1,7 @@
 package fr.maxlego08.menu.common.utils;
 
 import fr.maxlego08.menu.api.configuration.Configuration;
-import fr.maxlego08.menu.placeholder.Placeholder;
+import fr.maxlego08.menu.api.placeholder.Placeholder;
 import org.bukkit.OfflinePlayer;
 
 import java.util.ArrayList;
