+++ src/main/java/fr/maxlego08/menu/zcore/ZPlugin.java
@@ -4,6 +4,8 @@ import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import fr.maxlego08.menu.ZMenuPlugin;
 import fr.maxlego08.menu.api.exceptions.ListenerNullException;
+import fr.maxlego08.menu.api.placeholder.LocalPlaceholder;
+import fr.maxlego08.menu.api.placeholder.Placeholder;
 import fr.maxlego08.menu.api.players.Data;
 import fr.maxlego08.menu.api.utils.EnumInventory;
 import fr.maxlego08.menu.command.VCommand;
@@ -11,8 +13,6 @@ import fr.maxlego08.menu.command.VCommandManager;
 import fr.maxlego08.menu.inventory.VInventory;
 import fr.maxlego08.menu.inventory.VInventoryManager;
 import fr.maxlego08.menu.listener.ListenerAdapter;
-import fr.maxlego08.menu.placeholder.LocalPlaceholder;
-import fr.maxlego08.menu.placeholder.Placeholder;
 import fr.maxlego08.menu.zcore.logger.Logger;
 import fr.maxlego08.menu.zcore.logger.Logger.LogType;
 import fr.maxlego08.menu.zcore.utils.gson.DataAdapter;
